Hacking in on a migration

+++ yandex_money/models.py
@@ -103,7 +103,7 @@ class Payment(models.Model):
     status = models.CharField(
         'Статус', max_length=16, choices=STATUS.CHOICES,
         default=STATUS.PROCESSED)
-    invoice_id = models.PositiveIntegerField(
+    invoice_id = models.BigIntegerField(
         'Номер транзакции оператора', blank=True, null=True)
     shop_amount = models.DecimalField(
         'Сумма полученная на р/с', max_digits=15, decimal_places=2, blank=True,
